Kulikowa_lesson_10_4

+++ OnlineShop/OnlineShopWebApp/Areas/Admin/Controllers/UserController.cs
@@ -3,8 +3,11 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using OnlineShop.Db.Models;
 using OnlineShop.Db.Repositories;
+using OnlineShopWebApp.Areas.Admin.Models;
 using OnlineShopWebApp.Helpers;
 using OnlineShopWebApp.Models;
+using System.Collections.Generic;
+using System.Data;
 using System.Linq;
 
 namespace OnlineShopWebApp.Areas.Admin.Controllers
@@ -14,27 +17,39 @@ namespace OnlineShopWebApp.Areas.Admin.Controllers
     public class UserController : Controller
     {
         private readonly UserManager<User> userManager;
-        public UserController(UserManager<User> userManager)
+        private readonly RoleManager<IdentityRole> roleManager;
+        public UserController(UserManager<User> userManager, RoleManager<IdentityRole> roleManager)
         {
             this.userManager = userManager;
+            this.roleManager = roleManager;
         }
 
         public IActionResult Index()
         {
             var users = userManager.Users.ToList();
-            return View(users.ToUserViewModels());
+            var usersVM = new List<UserViewModel>();
+            foreach (var user in users)
+            {
+                var userVM = GetRolesVM(user);
+                usersVM.Add(userVM);
+            }
+            return View(usersVM);
         }
 
         public IActionResult Details(string id)
         {
             var user = userManager.FindByIdAsync(id).Result;
-            var userVM = user.ToUserViewModel();
+            var userVM = GetRolesVM(user);
             return View(userVM);
         }
 
         public IActionResult Delete(string id)
         {
             var user = userManager.FindByIdAsync(id).Result;
+            if (user.UserName == "admin@gmail.com")
+            {
+                return RedirectToAction("Error");
+            }
             userManager.DeleteAsync(user).Wait();
             return RedirectToAction("Index");
         }
@@ -51,13 +66,13 @@ namespace OnlineShopWebApp.Areas.Admin.Controllers
             var user = userManager.FindByIdAsync(passwordVM.Id).Result;
 
             var checkOldPassword = userManager.CheckPasswordAsync(user, passwordVM.OldPassword).Result;
-            if(!checkOldPassword)
+            if (!checkOldPassword)
             {
                 ModelState.AddModelError("", "Вы неверно ввели старый пароль!");
             }
 
             var checkNew = userManager.CheckPasswordAsync(user, passwordVM.NewPassword).Result;
-            if(checkNew)
+            if (checkNew)
             {
                 ModelState.AddModelError("", "Пароль не должен быть идентичен старому!");
             }
@@ -110,7 +125,56 @@ namespace OnlineShopWebApp.Areas.Admin.Controllers
                     return View("Edit", userVM);
                 }
             }
+            userVM = GetRolesVM(user);
             return View("Details", userVM);
+        }
+
+        public IActionResult Rights(string id)
+        {
+            var user = userManager.FindByIdAsync(id).Result;
+            var roles = userManager.GetRolesAsync(user).Result;
+            var rolesVM = roles.Select(x => new RoleViewModel { Name = x }).ToList();
+
+            var allRoles = roleManager.Roles.ToList();
+            var allRolesVM = allRoles.Select(x => x.ToRoleViewModel()).ToList();
+
+            var rightsVM = user.ToRightsViewModel(rolesVM, allRolesVM);
+            return View(rightsVM);
+        }
+
+        [HttpPost]
+        public IActionResult Rights(string id, Dictionary<string, string> userRolesVM)
+        {
+            var userSelectedRoles = userRolesVM.Select(x => x.Key);
+
+            var user = userManager.FindByIdAsync(id).Result;
+            var roles = userManager.GetRolesAsync(user).Result;
+
+            userManager.RemoveFromRolesAsync(user, roles).Wait();
+            try
+            {
+                userManager.AddToRolesAsync(user, userSelectedRoles).Wait();
+            }
+            catch
+            {
+                userManager.AddToRolesAsync(user, roles).Wait();
+            }
+
+            var userVM = GetRolesVM(user);
+            return View("Details", userVM);
+        }
+
+        public UserViewModel GetRolesVM(User user)
+        {
+            var roles = userManager.GetRolesAsync(user).Result.ToList();
+            var userVM = user.ToUserViewModel();
+            userVM.Roles = roles;
+            return userVM;
+        }
+
+        public IActionResult Error()
+        {
+            return View();
         }
     }
 }
